Update: quantity field type

+++ app/models/order.py
@@ -1,5 +1,5 @@
 from app.models.state import OrderStatus
-from sqlalchemy import Column, String, Enum, Integer, Date
+from sqlalchemy import Column, String, Enum, Integer, Date, Float
 from app.db.database import Base
 
 
@@ -7,15 +7,15 @@ class Order(Base):
     __tablename__ = "order"
     lote = Column(Integer, primary_key=True, nullable=False)
     description = Column(String)
-    quantity = Column(Integer)
+    quantity = Column(Float)
     dueDate = Column(Date)
     status = Column(Enum(OrderStatus), default=OrderStatus.unprogrammed)
     code = Column(String)
     bin = Column(Integer)
     received_user = Column(String)
     received_date = Column(Date)
-    received_quantity = Column(Integer)
-    missing_quantity = Column(Integer)
+    received_quantity = Column(Float)
+    missing_quantity = Column(Float)
     submitted_user = Column(String)
     submitted_date = Column(Date)
     submitted_observations = Column(String)
